[Prefabs] Reorder files to match concepts (#547)
* Reorder files to match concepts

* Put every prefab in root to avoid having a useless prefix

* Fix bad import in bunny example

* Update examples/stlib/PrefabScene_beginner.py


---------
