+++ examples/stlib/PrefabScene_beginner.py
@@ -1,7 +1,7 @@
-from stlib.entities.rigid import Rigid
-from stlib.entities.deformable import Deformable
-from stlib.geometry.cube import CubeParameters
-from stlib.geometry.file import FileParameters
+from stlib.materials.rigid import Rigid
+from stlib.materials.deformable import Deformable
+from stlib.geometries.cube import CubeParameters
+from stlib.geometries.file import FileParameters
 from splib.simulation.headers import setupLagrangianCollision
 from splib.simulation.linear_solvers import addLinearSolver
 from splib.simulation.ode_solvers import addImplicitODE
