+++ stlib/entities/__entity__.py
@@ -1,15 +1,12 @@
 from stlib.core.baseParameters import BaseParameters
-from stlib.core.basePrefab import BasePrefab
-from stlib.prefabs.collision import CollisionParameters, Collision
-from stlib.prefabs.visual import VisualParameters, Visual
-from stlib.prefabs.material import Material, MaterialParameters
-from stlib.geometry import Geometry
-from stlib.geometry.extract import ExtractParameters
+from stlib.collision import CollisionParameters, Collision
+from stlib.visual import VisualParameters, Visual
+from stlib.materials import Material, MaterialParameters
+from stlib.geometries import Geometry
 import dataclasses
-from typing import Callable, Optional, overload, Any
-from stlib.geometry import GeometryParameters
+from typing import Callable, Optional
+from stlib.geometries import GeometryParameters
 from splib.core.enum_types import StateType
-import Sofa
 from stlib.core.basePrefab import BasePrefab
