--- stlib/entities/deformable/__init__.py
@@ -1,2 +1,0 @@
-from  .__deformable__ import *
-from  .__parameters__ import *
