--- stlib/entities/rigid/__init__.py
@@ -1,1 +1,0 @@
-from .__rigid__ import *
