+++ stlib/geometries/__geometry__.py
