+++ stlib/geometries/__init__.py
