+++ stlib/geometries/cube.py
@@ -1,4 +1,4 @@
-from stlib.geometry import GeometryParameters
+from stlib.geometries import GeometryParameters
 
 class CubeParameters(GeometryParameters):
     def __init__(self,  center, edgeLength, pointPerEdge, dynamicTopology = False):
