+++ stlib/materials/__material__.py
@@ -11,7 +11,7 @@ class MaterialParameters(BaseParameters):
 
     massDensity : float = DEFAULT_VALUE
     massLumping : bool = DEFAULT_VALUE
-    
+
     stateType : StateType = StateType.VEC3
 
     addMaterial : Optional[Callable] = lambda node : addMass(node, node.parameters.stateType, massDensity=node.parameters.massDensity, lumping=node.parameters.massLumping)
@@ -24,7 +24,7 @@ class Material(BasePrefab):
 
     def __init__(self, parameters: MaterialParameters):
         BasePrefab.__init__(self, parameters)
-        
+
 
     def init(self):
         self.addObject("MechanicalObject", name="States", template=str(self.parameters.stateType))
