+++ stlib/materials/deformable.py
@@ -1,5 +1,5 @@
-from stlib.prefabs.material import MaterialParameters
-from splib.core.enum_types import ConstitutiveLaw, ElementType
+from stlib.materials import MaterialParameters
+from splib.core.enum_types import ConstitutiveLaw
 from stlib.core.baseParameters import Callable, Optional, dataclasses
 from splib.mechanics.linear_elasticity import *
 from splib.mechanics.hyperelasticity import *
@@ -26,9 +26,8 @@ class DeformableBehaviorParameters(MaterialParameters):
 
 def createScene(root) :
     from stlib.entities import Entity, EntityParameters
-    from stlib.prefabs.visual import VisualParameters
-    from stlib.geometry.extract import ExtractParameters
-    from stlib.geometry.file import FileParameters    
+    from stlib.visual import VisualParameters
+    from stlib.geometries.file import FileParameters
 
     root.addObject("VisualStyle", displayFlags=["showBehavior"])
