+++ stlib/visual.py
@@ -1,7 +1,7 @@
 from stlib.core.basePrefab import BasePrefab
-from stlib.core.baseParameters import BaseParameters, Optional, dataclasses, Any
-from stlib.geometry import Geometry, GeometryParameters
-from stlib.geometry.file import FileParameters
+from stlib.core.baseParameters import BaseParameters, Optional, dataclasses
+from stlib.geometries import Geometry, GeometryParameters
+from stlib.geometries.file import FileParameters
 from splib.core.utils import DEFAULT_VALUE
 from Sofa.Core import Object
